API/Reference

For now, here is a list of all documented functions.

Index

CCDReduction.combine

CCDReduction.crop

CCDReduction.cropview

CCDReduction.flat_correct

CCDReduction.flat_correct!

CCDReduction.subtract_bias

CCDReduction.subtract_bias!

CCDReduction.subtract_overscan

CCDReduction.subtract_overscan!

CCDReduction.trim

CCDReduction.trimview

API/Reference

CCDReduction.combine — Method

combine(frames::Vararg{<:AbstractArray}; method = median)
combine(frames::AbstractVector{<:AbstractArray}; method = median)

Combine multiple frames using method.

To pass a custom method, it must have a signature like method(::AbstractArray; dims).

Examples

julia> frame = [reshape(1.0:4.0, (2, 2)) for i = 1:4];

julia> combine(frame)
2×2 Array{Float64,2}:
 1.0  3.0
 2.0  4.0

julia> combine(frame, method = sum)
2×2 Array{Float64,2}:
 4.0  12.0
 8.0  16.0

CCDReduction.crop — Method

crop(frame::AbstractArray, shape; force_equal = true)

Crops frame to the size specified by shape anchored by the frame center.

This will remove rows/cols of the frame equally on each side. When there is an uneven difference in sizes (e.g. size 9 -> 6 can't be removed equally) the default is to increase the output size (e.g. 6 -> 7) so there is equal removal on each side. To disable this, set force_equal=false, which will remove the extra slice from the end of the axis.

Examples

julia> frame = reshape(1:25, (5, 5));

julia> crop(frame, (3, 3))
3×3 Array{Int64,2}:
 7  12  17
 8  13  18
 9  14  19

julia> crop(frame, (4, 3), force_equal = false)
4×3 Array{Int64,2}:
 6  11  16
 7  12  17
 8  13  18
 9  14  19

See Also

cropview

CCDReduction.cropview — Method

cropview(frame::AbstractArray, shape; force_equal = true)

Crops frame to the size specified by shape anchored by the frame center.

This function is same as the crop function but returns a view of the frame.

Note

This function returns a view of the frame, so any modification to output array will result in modification of frame.

See Also

crop

CCDReduction.flat_correct! — Method

flat_correct!(frame::AbstractArray, flat_frame::AbstractArray; norm_value = mean(flat_frame))

In-place version of flat_correct

See Also

flat_correct

CCDReduction.flat_correct — Method

flat_correct(frame::AbstractArray, flat_frame::AbstractArray; norm_value = mean(flat_frame))

Correct frame for non-uniformity using the calibrated flat_frame.

By default, the flat_frame is normalized by its mean, but this can be changed by providing a custom norm_value.

Note

This function may introduce non-finite values if flat_frame contains values very close to 0 due to dividing by zero. The default behavior will return Inf if the frame value is non-zero, and Nan if the frame value is 0.

Examples

julia> frame = ones(3, 3);

julia> flat = fill(2.0, (3, 3));

julia> flat_correct(frame, flat, norm_value = 1.0)
3×3 Array{Float64,2}:
 0.5  0.5  0.5
 0.5  0.5  0.5
 0.5  0.5  0.5

julia> flat_correct(frame, flat)
3×3 Array{Float64,2}:
 1.0  1.0  1.0
 1.0  1.0  1.0
 1.0  1.0  1.0

See Also

flat_correct!

CCDReduction.subtract_bias! — Method

subtract_bias!(frame::AbstractArray, bias_frame::AbstractArray)

In-place version of subtract_bias

See Also

subtract_bias

CCDReduction.subtract_bias — Method

subtract_bias(frame::AbstractArray, bias_frame::AbstractArray)

Subtract the bias_frame from frame.

Examples

julia> frame = [1.0 2.2 3.3 4.5];

julia> bias = [0.0 0.2 0.3 0.5];

julia> subtract_bias(frame, bias)
1×4 Array{Float64,2}:
 1.0  2.0  3.0  4.0

See Also

subtract_bias!

CCDReduction.subtract_overscan! — Method

subtract_overscan!(frame::AbstractArray, idxs; dims = axes_min_length(idxs))

In-place version of subtract_overscan

See Also

subtract_overscan

CCDReduction.subtract_overscan — Method

subtract_overscan(frame::AbstractArray, idxs; dims = axes_min_length(idxs))

Subtract the overscan frame from image.

dims is the dimension along which overscan_frame is combined. The default value of dims is the axis with smaller length in overscan region. If idxs is a string it will be parsed as FITS-style indices.

Examples

julia> frame = [4.0 2.0 3.0 1.0 1.0];

julia> subtract_overscan(frame, (:, 4:5), dims = 2)
1×5 Array{Float64,2}:
 3.0  1.0  2.0  0.0  0.0

julia> subtract_overscan(frame, "[4:5, 1:1]", dims = 2)
1×5 Array{Float64,2}:
 3.0  1.0  2.0  0.0  0.0

See Also

subtract_overscan!

CCDReduction.trim — Method

trim(frame::AbstractArray, idxs)

Trims the frame to remove the region specified by idxs.

This function trims the array in a manner such that final array should be rectangular. The indices follow standard Julia convention, so (:, 45:60) trims all columns from 45 to 60 and (1:20, :) trims all the rows from 1 to 20. The function also supports FITS-style indices.

Examples

julia> frame = ones(5, 5);

julia> trim(frame, (:, 2:5))
5×1 Array{Float64,2}:
 1.0
 1.0
 1.0
 1.0
 1.0

julia> trim(frame, "[2:5, 1:5]")
5×1 Array{Float64,2}:
 1.0
 1.0
 1.0
 1.0
 1.0

See Also

trimview

CCDReduction.trimview — Method

trimview(frame::AbstractArray, idxs)

Trims the frame to remove the region specified by idxs.

This function is same as the trim function but returns a view of the frame.

Note

This function returns a view of the frame, so any modification to output array will result in modification of frame.

See Also

trim